requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
boto3==1.34.162
botocore==1.34.162
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
docker-pycreds==0.4.0
easydict==1.13
efficientnet_pytorch==0.7.1
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.35.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.4
jmespath==1.0.1
keras==3.4.1
kiwisolver==1.4.5
lapx==0.5.5
lazy_loader==0.4
libclang==18.1.1
lu-vp-detect==1.0.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
mediapipe==0.10.14
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
munch==4.0.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
nvcc4jupyter==1.0.0
opencv-contrib-python==4.10.0.84
opencv-contrib-python-headless==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pretrainedmodels==0.7.4
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
segmentation-models-pytorch==0.3.3
sentry-sdk==2.13.0
setproctitle==1.3.3
setuptools==72.1.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sounddevice==0.5.0
sympy==1.13.1
tbb==2021.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.8.30
timm==0.9.2
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.91
ultralytics-thop==2.0.0
urllib3==2.2.2
uvicorn==0.30.6
wandb==0.17.7
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
